% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabel.R
\name{addLabel}
\alias{addLabel}
\title{Add label to Raman plot}
\usage{
addLabel(data, x, label = NULL)
}
\arguments{
\item{data}{raman data. accept both matrix and hyperSpec class.}

\item{x}{x value for labels to appear}

\item{label}{labels. If not supplied will use sample names in data.}
}
\description{
Add label to Raman plot
}
\examples{
data(raman)
plotRaman(raman)
addLabel(raman, 1300)

}

