% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addstats.R
\name{addstats}
\alias{addstats}
\title{Add stats of the model to the plot}
\usage{
addstats(model, ncomp = length(model$ncomp), estimate = "CV",
  legendlocation = "bottomright", show = c("R2", "RMSECV", "RMSEC"))
}
\arguments{
\item{model}{object of class `mvr`}

\item{ncomp}{a vector of positive integers.  The components or number
    of components to use.  See below.}

\item{estimate}{a character vector.  Which estimators to use.
    Should be a subset of \code{c("all", "train", "CV", "adjCV",
      "test")}.  \code{"adjCV"} is only available for (R)MSEP.  See
    below for how the estimators are chosen.}

\item{legendlocation}{location of legend on graph. Look up legend for more details.}
}
\description{
Add stats of the model to the plot
}

