% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageEEM.R
\name{averageEEM}
\alias{averageEEM}
\title{Calculate average of EEM}
\usage{
averageEEM(EEM, use = "mean")
}
\arguments{
\item{use}{`mean` or `median`}

\item{EEM_uf}{unfolded EEM}
}
\description{
Calculate average of EEM
}

