% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biplot4.R
\name{biplot4}
\alias{biplot4}
\title{Draw scores as points}
\usage{
biplot4(x, ...)
}
\arguments{
\item{legendoutside}{(optional) set to TRUE if you want to put legend on
the outside of the plot. The legend location is defaulted to topright.}

\item{rightwhitespace}{(optional) set width for white space for legend.
Only applicable if legendoutside = TRUE}

\item{legendinset}{(optional) how much legend box will be pushed to the right.
Only applicable if legendoutside = TRUE}
}
\description{
I'm too lazy to fix the code so please always specify xcol for this to work.
Biplot4 is similar to biplot3 except that the legend will be place outside
}
\examples{
require(EEM)
data(applejuice)
applejuice_uf <- unfold(applejuice) # unfold list into matrix
# get country of apple production
country <- sapply(strsplit(names(applejuice), split = "-"), "[", 1)

# select peaks
local_peak <- findLocalMax(applejuice, n = 1)
index <- colnames(applejuice_uf) \%in\% local_peak
applejuice_uf_selectedPeak <- applejuice_uf[,index, drop = FALSE]

# PCA
result <- prcomp(applejuice_uf_selectedPeak)

# create color palette for x points
library(RColorBrewer)
xcol <- brewer.pal(3, "Dark2")
biplot4(result, xlab = prcompname(result ,1), ylab = prcompname(result,2),
xlabs = country, xcol = xcol)
}

