% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deflate.R
\name{deflate}
\alias{deflate}
\alias{deflate.character}
\alias{deflate.matrix}
\title{Deflate}
\usage{
deflate(x, n, ...)

\method{deflate}{matrix}(x, n, process = "average", sdigits = 2,
  rowname = NULL)

\method{deflate}{character}(x, n, FUN, ...)
}
\arguments{
\item{x}{object}

\item{n}{number of rows per subset to average}

\item{...}{arguments to pass to function}

\item{process}{processing method. Available option: none, average, std}

\item{sdigits}{significant figures}

\item{FUN}{function}
}
\description{
Deflate
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: for matrix

\item \code{character}: for character
}}

