% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delZeroCol.R
\name{delZeroCol}
\alias{delZeroCol}
\title{Delete columns filled with zero}
\usage{
delZeroCol(uf)
}
\arguments{
\item{uf}{unfolded matrix}
}
\description{
Delete columns filled with zero
}
\examples{
require(EEM)
data(applejuice)
applejuice_uf <- unfold(applejuice)
dim(applejuice_uf)
applejuice_uf_nozero <- delZeroCol(applejuice_uf)
dim(applejuice_uf_nozero)
}

