% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawEEMgg.R
\name{drawEEMgg}
\alias{drawEEMgg}
\alias{drawEEMgg.EEM}
\alias{drawEEMgg.EEMweight}
\title{Draw contour for EEM data using ggplot2}
\usage{
drawEEMgg(x, ...)

\method{drawEEMgg}{EEM}(x, n, textsize = 25, color = matlab.like,
  geom = "path", nlevel = 20, xlab = "Excitation wavelength [nm]",
  ylab = "Emission wavelength [nm]", title = NULL, has_legend = TRUE,
  zlim = NULL, breaks = waiver())

\method{drawEEMgg}{EEMweight}(x, ncomp, textsize = 25, color = matlab.like,
  geom = "path", nlevel = 20, xlab = "Excitation wavelength [nm]",
  ylab = "Emission wavelength [nm]", title = NULL, has_legend = TRUE,
  zlim = NULL, breaks = waiver())
}
\arguments{
\item{textsize}{(optional) text size}

\item{geom}{(optional) object to display data. Can be "polygon" or "path" but default to "path" at the moment to
due to some bugs for "polygon".}
}
\value{
A figure is returned on the graphic device
}
\description{
This function draw contour for EEM data using ggplot2. Use `ggsave` to save the contours.
}
\details{
\code{\link{drawEEM}} is faster and should be used. But since it lacks multiple plotting capacity, \code{\link{drawEEMgg}} was created to work with drawMultipleEEM.
}
\section{Methods (by class)}{
\itemize{
\item \code{EEM}: draw EEM of EEM data created by \code{\link{readEEM}} function

\item \code{EEMweight}: draw contours of the output from \code{\link[EEM]{getLoading}} and
\code{\link[EEM]{getReg}}.
}}
\examples{
\dontrun{
require(EEM)
data(applejuice)
drawEEMgg(applejuice, 1) # draw EEM of sample no.1
drawEEMgg(applejuice, 1, color = cm.colors) # draw EEM of sample no.31 with different color
drawEEMgg(applejuice, 1, nlevel = 5) # reduce nlevel

# can be combined with other ggplot2 commands
drawEEMgg(applejuice, 1) + theme_bw()
}
}
\seealso{
\code{\link{drawEEM}},\code{\link{drawMultipleEEM}}
}

