% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMultipleEEM.R
\name{drawMultipleEEM}
\alias{drawMultipleEEM}
\title{Draw multiple contours for EEM data}
\usage{
drawMultipleEEM(EEM, ggplot = TRUE, textsize = 13, ncol = NULL,
  nrow = NULL, ...)
}
\arguments{
\item{ncol}{An integer describing the number of columns in the layout.}

\item{nrow}{An integer describing the number of rows in the layout.}

\item{...}{(optional) arguments for \code{\link{drawEEMggplot}}}
}
\value{
Multiple figures returned on the graphic device
}
\description{
Draws multiple contours for EEM data using ggplot2 with the help of GridExtra package
}
\examples{
require(EEM)
data(applejuice)
drawMultipleEEM(applejuice[1:5]) # draw contour of the first to the fifth sample

}
\seealso{
\code{\link{drawEEM}},\code{\link{drawEEMggplot}}
}

