% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSpec.R
\name{drawSpec}
\alias{drawSpec}
\title{Draw spectra}
\usage{
drawSpec(x, ...)
}
\arguments{
\item{x}{unfolded EEM or EEM}

\item{EX}{excitation wavelength}

\item{EM}{emission wavelength}

\item{group}{a vector of characters or factors}

\item{ggplot}{logical value whether to use ggplot or not}

\item{legendlocation}{legend location. can be anything from "bottomright", 
"bottom", "bottomleft", "left", "topleft", "top", "topright", "right" or "center".}
}
\description{
Draw spectra
}
\examples{
require(EEM)
data(applejuice)
country <- sapply(strsplit(names(applejuice), split = "-"), "[", 1)

# ggplot
drawSpec(unfold(applejuice), EX = 340)
drawSpec(unfold(applejuice), EX = 340, group = country)
drawSpec(unfold(applejuice), EM = 400, group = country)

# base plot
drawSpec(unfold(applejuice), EX = 340, group = country, ggplot = FALSE)
drawSpec(unfold(applejuice), EM = 400, group = country, ggplot = FALSE)

}

