% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcProcessEEM.R
\name{gcProcessEEM}
\alias{gcProcessEEM}
\title{Group and column-wise process EEM data matrix}
\usage{
gcProcessEEM(EEM, group, FUN)
}
\arguments{
\item{EEM}{EEM object}

\item{group}{a character or factor vector}

\item{FUN}{function to apply}
}
\description{
Group and column-wise process EEM data matrix
}
\examples{
require(EEM)
data(applejuice)
applejuice_mean <- gcProcessEEM(applejuice, country, mean)
drawEEM(applejuice_mean, 1)

}

