% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR2.R
\name{getR2}
\alias{getR2}
\title{get R2 from mvr object}
\usage{
getR2(mvr, ncomp = mvr$ncomp, estimate, newx = NULL, newy = NULL,
  showprint = TRUE)
}
\arguments{
\item{mvr}{mvr object}

\item{ncomp}{number of component}

\item{estimate}{a subset of c("train", "CV", "test")}

\item{newx}{predictors for test group}

\item{newy}{measured value for test group}

\item{showprint}{show result in console}
}
\description{
Alternative to pls package's `R2`. While `R2` requires a `newdata` dataframe which combines
both predictors and target, `getR2` lets you put in `newx` and `newy` separately. 
Aside from that, declaring `estimate` in `getR2` remind you which value you got.
}
\examples{
require(pls)
data(yarn)
model <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
R2(model)
getR2(model, estimate = "train") # return R2 at particular ncomp without intercept value
getR2(model, estimate = "CV")

}

