% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRMSE.R
\name{getRMSE}
\alias{getRMSE}
\title{get RMSE}
\usage{
getRMSE(mvr, ncomp = mvr$ncomp, estimate, newx = NULL, newy = NULL)
}
\arguments{
\item{mvr}{mvr object}

\item{estimate}{a subset of c("train", "CV", "test")}

\item{newx}{predictors for test group}

\item{newy}{measured value for test group}
}
\description{
get RMSE
}

