% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRaman.R
\name{plotRaman}
\alias{plotRaman}
\title{Plot raman spectra}
\usage{
plotRaman(data, group = NULL, legendlocation = "topright",
  xlab = "Raman shift (1/cm)", ylab = "Counts", ...)
}
\arguments{
\item{data}{matrix or hyperSpec class object}

\item{group}{a vector of characters or factors describing categories}

\item{legendlocation}{location of legend. can be anything from "bottomright",
"bottom", "bottomleft", "left", "topleft", "top", "topright", "right" or "center".}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{extra arguments passing to plotspc or matplot}
}
\description{
Plot raman spectra
}

