% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRamanLoading.R
\name{plotRamanLoading}
\alias{plotRamanLoading}
\title{Plot PCA loadings for Raman}
\usage{
plotRamanLoading(PCA, ncomp = 1:6, xlab = "Raman shift (1/cm)",
  ylab = "loading", ...)
}
\arguments{
\item{PCA}{output variable from \code{\link[stats]{prcomp}}}

\item{ncomp}{a vector of number of components}

\item{ylab}{y-axis label}

\item{...}{additional arguments for \code{\link[graphic]{matplot}}}
}
\description{
Plot PCA loadings for Raman
}
\examples{
data(raman)
PCA <- prcomp(raman)
plotRamanLoading(PCA)
plotRamanLoading(PCA, ncomp = 1:2) # display only the first two components
plotRamanLoading(PCA, ncomp = c(1,3)) # display only the first and the third component

}

