% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScore_ellipse.R
\name{plotScore_ellipse}
\alias{plotScore_ellipse}
\title{Plot PCA scores with ellipse}
\usage{
plotScore_ellipse(PCA, x = 1, y = 2, level = 0.95, xlim = NULL,
  ylim = NULL, ...)
}
\description{
Plot PCA scores with ellipse using \code{\link{dataEllipse}}
}
\examples{
require(EEM)
data(applejuice)
applejuice_uf <- unfold(applejuice)
PCA <- prcomp(applejuice_uf)
plotScore_ellipse(PCA)

# change level
plotScore_ellipse(PCA, level = 0.8) 

# manually set x,y ranges
plotScore_ellipse(PCA, xlim = c(-9000, 9000), ylim = c(-4000, 4000)) 

# fill in circles
plotScore_ellipse(PCA, fill = TRUE, 
fill.alpha = 0.2, xlim = c(-9000, 9000), ylim = c(-4000, 4000))


}

