% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVIP.R
\name{plotVIP}
\alias{plotVIP}
\title{Plot VIP for EEM data}
\usage{
plotVIP(x, ncomp = NULL, ...)
}
\arguments{
\item{x}{output variable from \code{\link[pls]{plsr}} function}

\item{ncomp}{number of components/ latent variables}

\item{...}{(optional) arguments for \code{\link[EEM]{drawEEM}} and \code{\link[graphics]{filled.contour}}}
}
\value{
A figure is returned on the graphic device
}
\description{
Plot VIP for EEM data
}
\examples{
require(EEM)
data(gluten)
gluten_uf <- unfold(gluten) # unfold list into matrix

# delete columns with NA values
index <- colSums(is.na(gluten_uf)) == 0
gluten_uf <- gluten_uf[, index]
gluten_ratio <- as.numeric(names(gluten))

require(pls)
model <- plsr(gluten_ratio ~ gluten_uf, ncomp = 3, method = "oscorespls") 
plotVIP(model) 
 
}

