% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsplot.R
\name{plsplot}
\alias{plsplot}
\title{Plot prediction graph for mvr model}
\usage{
plsplot(mvr, ncomp = mvr$ncomp, estimate = "train",
  location = "bottomright", show = c("ncomp", "R2", "RMSE"), round = 2,
  newx = NULL, newy = NULL, fitline = TRUE, cex.stats = 1, ...)
}
\arguments{
\item{mvr}{mvr object (pls model)}

\item{ncomp}{number of component}

\item{location}{legend position}

\item{show}{select statistics to show}

\item{round}{digits to round}

\item{newx}{newx}

\item{newy}{newy}

\item{fitline}{add fitline or not}

\item{cex.stats}{text size of statistics legend}

\item{...}{pass on to predplot}
}
\description{
Plot prediction graph using pls package's `predplot` function but with ability 
to add statistical labeling
}
\examples{
require(pls)
data(yarn)
model <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
plsplot(model) # calibration set result
plsplot(model, estimate = "CV") # cross validation set result

## customizing the graphs
plsplot(model, location = "topleft") # change legend position
plsplot(model, show = "R2") # show only R2
plsplot(model, show = "R2", round = 4) # round to four digits
plsplot(model, fitline = FALSE) # get rid of fitline
plsplot(model, show = "R2", cex.stats = 3) # bigger stats font
plsplot(model, cex.lab = 1.5, cex.main = 2) # bigger labels font


}

