% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plsplot2.R
\name{plsplot2}
\alias{plsplot2}
\title{plot both calibration and validation groups on the same graphs}
\usage{
plsplot2(model, newx = NULL, newy = NULL, ncomp = model$ncomp,
  location = "bottomright", show = c("ncomp", "R2", "RMSE"), round = 2,
  fitline = TRUE, cex.pt = 1.5, cex.stats = 1, xlab = "Measured value",
  ylab = "Predicted value", col.cal = "black", col.val = "grey",
  pch.cal = 21, pch.val = 21, ...)
}
\description{
plot both calibration and validation groups on the same graphs
}

