% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trainPLS.R
\name{predict.trainPLS}
\alias{predict.trainPLS}
\title{predict function for trainPLS object}
\usage{
\method{predict}{trainPLS}(object, newx, ...)
}
\arguments{
\item{...}{arguments for \code{predict} function}

\item{x}{trainPLS object}
}
\description{
predict function for trainPLS object
}

