% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_trainPLS.R
\name{print.trainPLS}
\alias{print.trainPLS}
\title{Print trainPLS}
\usage{
\method{print}{trainPLS}(x, ...)
}
\arguments{
\item{x}{trainPLS object}

\item{...}{arguments for \code{print} function}
}
\description{
Print trainPLS
}

