% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readRaman.R
\name{readRaman}
\alias{readRaman}
\title{Read raw Raman files}
\usage{
readRaman(path, hyperSpec = FALSE)
}
\arguments{
\item{path}{path to the files or folders which contains raw files}
}
\description{
Read raw Raman files
}
\examples{
# not run
# data <- readRaman(getwd()) # data in working directory
# data <- readRaman("folder") # data in folder
# data <- readRaman("file.txt") # specify file
# data[1:5,1:5]
}

