% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveContour.R
\name{saveContour}
\alias{saveContour}
\title{Export all contours in FF object to png graphics}
\usage{
saveContour(FF, path = NULL, size = c("normal", "big"), rep = 1,
  width = NULL, height = NULL, pointsize = NULL, cex.lab = 1.3,
  cex.main = 1.3, ...)
}
\description{
Export all contours in FF object to png graphics
}

