% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR.R
\name{SR}
\alias{SR}
\title{Calculate SR from mvr}
\usage{
SR(mvr)
}
\arguments{
\item{mvr}{pls model from pls package}

\item{ncomp}{ncomp}
}
\description{
Mostly edited and made to work by Y. Uwadaira of \code{plsropt} package.
Validated with the values from MATLAB's PLS toolbox v8.1.1.
}

