% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainPLS.R
\name{trainPLS}
\alias{trainPLS}
\title{Train PLS for train dataset by cross-validation}
\usage{
trainPLS(x, y, newx = NULL, newy = NULL, maxncomp = 20, cvsegments = 10,
  round = 2)
}
\arguments{
\item{cvsegments}{refer to mvrCv's segments argument}
}
\description{
Train PLS for train dataset by cross-validation. The preprocessing method will be optimized automatically.
However, the number of latent variables has to be determined manually. Planning to add variable reduction in the future.
}

