% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainPLS2.R
\name{trainPLS2}
\alias{trainPLS2}
\title{Train PLS for train dataset by cross-validation}
\usage{
trainPLS2(x, y, newx = NULL, newy = NULL, maxncomp = 20,
  cvsegments = 10, round = 2, reduceVar = TRUE, cycles = 3,
  fixedncomp = NULL, prepro = c("mc", "norm_mc", "au"))
}
\arguments{
\item{cvsegments}{refer to mvrCv's segments argument}

\item{fixedncomp}{fixed numerical value}
}
\description{
Train PLS for train dataset by cross-validation. The preprocessing method will be optimized automatically.
However, the number of latent variables has to be determined manually. This is different from trainPLS as
you have to specify the preprocessing method manually.
}

