% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPLS_general.R
\name{trainPLS_general}
\alias{trainPLS_general}
\title{Train PLS for train dataset by cross-validation}
\usage{
trainPLS_general(x, y, maxncomp = 20, cvsegments = 10, round = 2,
  ncomp = c("auto", "manual", "fixed"), fixedncomp = NULL,
  threshold = 0.02, saveModel = FALSE, plotting = TRUE)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{prediction target vector}

\item{maxncomp}{maximum ncomp for calculation}

\item{cvsegments}{refer to mvrCv's segments argument}

\item{ncomp}{`auto`,`manual` or `fixed`}

\item{fixedncomp}{fixed numerical value}

\item{threshold}{threshold for selecting ncomp}
}
\description{
Train PLS for train dataset by cross-validation. The preprocessing method will be optimized automatically.
However, the number of latent variables has to be determined manually. Planning to add variable reduction in the future.
}

