% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deflate.R
\name{deflate}
\alias{deflate}
\title{Deflate}
\usage{
deflate(x, n, ...)
}
\arguments{
\item{x}{object}

\item{n}{number of rows per subset to average}

\item{...}{arguments to pass to function}

\item{process}{processing method. Available option: none, average, std}

\item{sdigits}{significant figures}

\item{FUN}{function}
}
\description{
Deflate
}

