% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawEEMgg.R
\name{drawEEMgg}
\alias{drawEEMgg}
\alias{drawEEMgg.EEM}
\alias{drawEEMgg.EEMweight}
\title{Draw contour for EEM data using ggplot2}
\usage{
drawEEMgg(x, ...)

\method{drawEEMgg}{EEM}(x, n, textsize = 20, color = matlab.like,
  geom = "polygon", nlevel = 50, xlab = "EX [nm]", ylab = "EM [nm]",
  title = NULL)

\method{drawEEMgg}{EEMweight}(x, ncomp, textsize = 20, color = matlab.like,
  geom = "polygon", nlevel = 50, xlab = "EX [nm]", ylab = "EM [nm]",
  title = NULL)
}
\arguments{
\item{textsize}{(optional) text size}

\item{geom}{(optional) object to display data. Can be "polygon" or "path".}
}
\value{
A figure is returned on the graphic device
}
\description{
This function draw contour for EEM data using ggplot2.
}
\details{
\code{\link{drawEEM}} is faster and should be used. But since it lacks multiple plotting capacity, \code{\link{drawEEMgg}} was created to work with \code{\link{drawMultipleEEM}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{EEM}: draw EEM of EEM data created by \code{\link{readEEM}} function

\item \code{EEMweight}: draw contours of the output from \code{\link[EEM]{getLoading}} and
\code{\link[EEM]{getReg}}.
}}
\examples{
data(AppleJuice)
drawEEMgg(AppleJuice, n = 36) # draw EEM of sample no.36
drawEEMgg(AppleJuice, n = 36, color = cm.colors) # draw EEM of sample no.36 with different color
drawEEMgg(AppleJuice, n = 36, geom = "path", nlevel = 20) # draw EEM of sample no.36 with different color
}
\seealso{
\code{\link{drawEEM}},\code{\link{drawMultipleEEM}}
}

