% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawSpec.R
\name{drawSpec}
\alias{drawSpec}
\title{Draw spectra}
\usage{
drawSpec(EEM_uf, EX = NULL, EM = NULL, group = NULL, ggplot = TRUE,
  legendlocation = "topright")
}
\description{
Draw spectra
}
\examples{
require(EEM)
data(applejuice)
country <- sapply(strsplit(names(applejuice), split = "-"), "[", 1)

# ggplot
drawSpec(unfold(applejuice), EX = 340)
drawSpec(unfold(applejuice), EX = 340, group = country)
drawSpec(unfold(applejuice), EM = 400, group = country)

# base plot
drawSpec(unfold(applejuice), EX = 340, group = country, ggplot = FALSE)
drawSpec(unfold(applejuice), EM = 400, group = country, ggplot = FALSE)
}

