% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRaman.R
\name{plotRaman}
\alias{plotRaman}
\title{Plot raman spectra}
\usage{
plotRaman(data, group = NULL, legendlocation = "topright", ...)
}
\arguments{
\item{group}{a vector of characters or factors describing categories}

\item{...}{extra arguments passing to matplot}

\item{hyperSpec}{hyperSpec class object}
}
\description{
Plot raman spectra
}

