% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRamanLoading.R
\name{plotRamanLoading}
\alias{plotRamanLoading}
\title{Plot PCA loadings for Raman}
\usage{
plotRamanLoading(PCA, ncomp = 1:6, ylab = "loading", ...)
}
\arguments{
\item{PCA}{output variable from \code{\link[stats]{prcomp}}}

\item{ncomp}{a vector of number of components}

\item{ylab}{y-axis label}

\item{...}{additional arguments for \code{\link[graphic]{matplot}}}
}
\description{
Plot PCA loadings for Raman
}

