% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameDimEEM.R
\name{sameDimEEM}
\alias{sameDimEEM}
\title{check that all EEM has the same dimension}
\usage{
sameDimEEM(EEM)
}
\arguments{
\item{EEM}{a list containing EEM data as created by \code{\link[EEM]{readEEM}} function}
}
\description{
check that all EEM has the same dimension
}

