% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainPLS2.R
\name{trainPLS2}
\alias{trainPLS2}
\title{Train PLS for train dataset by cross-validation}
\usage{
trainPLS2(x, y, newx = NULL, newy = NULL, maxncomp = 20,
  cvsegments = 10, round = 2, reduceVar = TRUE, cycles = 3,
  fixedncomp = NULL, prepro = c("mc", "norm_mc", "au"))
}
\arguments{
\item{x}{predictor matrix}

\item{y}{prediction target vector}

\item{maxncomp}{maximum ncomp for calculation}

\item{cvsegments}{refer to mvrCv's segments argument}

\item{round}{round numbers}

\item{reduceVar}{variable reduction using VIP}

\item{cycles}{cycles for variable reduction}

\item{fixedncomp}{fixed numerical value}

\item{prepro}{preprocessing method. Choose from c("mc", "norm_mc","au"). Default to "mc" if not specified.}
}
\description{
Train PLS for train dataset by cross-validation. This is different from trainPLS as
you have to specify the preprocessing method manually.
}

